backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
python-dotenv
db-dtypes
statsmodels
numpy
prophet
